+++ admin-frontend/open_admin_app/lib/widgets/features/cell-view/flag_colored_on_off_label.dart
