+++ admin-frontend/open_admin_app/lib/widgets/features/cell-view/strategy_tooltip.dart
