+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/boolean/edit_boolean_value_dropdown_widget.dart
