+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/individual_strategy_bloc.dart
