+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/lock_unlock_switch.dart
