+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/number/edit_number_value_container.dart
