+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/retire_feature_value_checkbox_widget.dart
