+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/attribute_value_chip_widget.dart
