+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/delete_strategy_icon_button.dart
