+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/split_rollout_button.dart
@@ -2,8 +2,8 @@ import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_icons/flutter_icons.dart';
 import 'package:mrapi/api.dart';
-import 'package:open_admin_app/widgets/features/custom_strategy_bloc.dart';import 'package:open_admin_app/widgets/features/table-expanded-view/individual_strategy_bloc.dart';
-import 'package:open_admin_app/widgets/features/table-expanded-view/strategies/strategy_editing_widget.dart';
+import 'package:open_admin_app/widgets/features/custom_strategy_bloc.dart';import 'package:open_admin_app/widgets/features/edit-feature-value/individual_strategy_bloc.dart';
+import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/strategy_editing_widget.dart';
 
 class AddStrategyButton extends StatelessWidget {
   final CustomStrategyBloc bloc;
