+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/strategy_card_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:mrapi/api.dart';
 import 'package:open_admin_app/widgets/features/custom_strategy_bloc.dart';import 'package:open_admin_app/widgets/features/feature_dashboard_constants.dart';
-import 'package:open_admin_app/widgets/features/table-expanded-view/strategies/delete_strategy_icon_button.dart';
-import 'package:open_admin_app/widgets/features/table-expanded-view/strategies/edit_value_strategy_link_button.dart';
+import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/delete_strategy_icon_button.dart';
+import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/edit_value_strategy_link_button.dart';
 
 class StrategyCardWidget extends StatelessWidget {
   final bool editable;
