+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/string_caps_extension.dart
