+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/transform_strategy_conditions.dart
