+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/string/edit_string_value_container.dart
